feat: set the final XVS legacy rewards to send to the vault
