+++ simulations/vip-230/simulations.ts
@@ -5,7 +5,7 @@ import { ethers } from "hardhat";
 
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import { vip228 } from "../../vips/vip-228";
+import { vip230 } from "../../vips/vip-230";
 import ERC20_ABI from "./abi/IERC20UpgradableAbi.json";
 import REWARD_FACET_ABI from "./abi/RewardFacet.json";
 
@@ -30,7 +30,7 @@ forking(34925177, () => {
     });
   });
 
-  testVip("VIP-228 Venus Recommend Parameters", vip228(), {
+  testVip("VIP-230 Venus Recommend Parameters", vip230(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [REWARD_FACET_ABI], ["VenusGranted"], [1]);
     },
@@ -38,11 +38,11 @@ forking(34925177, () => {
 
   describe("Post-VIP behavior", async () => {
     it("Check XVS Balance", async () => {
-      const _newBalanceComptroller = BigNumber.from("14303482694193242357204594").sub(parseUnits("96075", 18));
+      const _newBalanceComptroller = BigNumber.from("14303482694193242357204594").sub(parseUnits("191100", 18));
       const newBalanceComptroller = await xvs.balanceOf(COMPTROLLER);
       expect(_newBalanceComptroller).to.equal(newBalanceComptroller);
 
-      const _newBalanceXVSStore = BigNumber.from("43993251264814789517747").add(parseUnits("96075", 18));
+      const _newBalanceXVSStore = BigNumber.from("43993251264814789517747").add(parseUnits("191100", 18));
       const newBalanceXVSStore = await xvs.balanceOf(XVS_STORE);
       expect(_newBalanceXVSStore).to.equal(newBalanceXVSStore);
     });
